+++ packages/terra-framework-docs/src/terra-dev-site/test/date-picker/DatePickerStartDate.test.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import DatePicker from 'terra-date-picker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
+
+const DatePickerStartDate = () => (
+  <div className={cx('content-wrapper')}>
+    <DatePicker
+      name="date-input"
+      selectedDate="2017-04-01"
+    />
+  </div>
+);
+
+export default DatePickerStartDate;