+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerDefaultDateOnly.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerDefaultDateOnly = () => (
+  <DateTimePickerExampleTemplate
+    value="2017-08-08"
+  />
+);
+
+export default DateTimePickerDefaultDateOnly;