+++ packages/terra-framework-docs/src/terra-dev-site/test/date-time-picker/DateTimePickerInvalid.test.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import DateTimePickerExampleTemplate from './common/DateTimePickerExampleTemplate';
+
+const DateTimePickerInvalid = () => (
+  <DateTimePickerExampleTemplate
+    isInvalid
+  />
+);
+
+export default DateTimePickerInvalid;